用户模块后端
1、开发核心功能，用户注册、登录注销等
2、全局权限校验AOP
3、自定义配置解决精度丢失问题

+++ src/main/java/com/hl/hlaicodemother/common/PageRequest.java
@@ -15,12 +15,12 @@ public class PageRequest implements Serializable {
     /**
      * 当前页码，默认为1
      */
-    private int pageNum = 1;
+    private long pageNum = 1;
 
     /**
      * 每页条数，默认为10
      */
-    private int pageSize = 10;
+    private long pageSize = 10;
 
     /**
      * 排序字段，默认为空
